requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
alembic==1.14.0
pyppeteer==2.0.0
pyyaml==6.0.2
Flask-WTF==1.2.1
WTForms==3.1.2
